fix: 안심구역 설정 시 마커와 반경이 보이지 않는 오류 수정

+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/model/nok/safearea/RegisterSafeAreaRequest.kt
@@ -6,5 +6,5 @@ data class RegisterSafeAreaRequest(
     val areaName: String,
     val latitude: Double,
     val longitude: Double,
-    val radius: Int,
+    val radius: Float,
 )
